refactor(web): split use-toast into context/hook/provider modules
`use-toast.tsx` exported the Context, the hook, AND the provider
component from one file — same react-refresh/only-export-components
violation that `surface.tsx` had. Split analogously:

- hooks/toast-context.ts — Context object + types (no component)
- hooks/use-toast.ts      — `useToast()` consumer hook
- hooks/ToastProvider.tsx — provider component

Consumer import paths (`@/hooks/use-toast`) still resolve; only
main.tsx is updated for the new provider location. Lint now clean.
