+++ projects/web/src/hooks/ToastProvider.tsx
@@ -1,21 +1,5 @@
-import { useState, useCallback, createContext, useContext, type ReactNode } from 'react';
-
-export type ToastVariant = 'default' | 'destructive' | 'success';
-
-export interface Toast {
-  id: string;
-  title?: string;
-  description?: string;
-  variant?: ToastVariant;
-}
-
-interface ToastContextType {
-  toasts: Toast[];
-  toast: (props: Omit<Toast, 'id'>) => void;
-  dismiss: (id: string) => void;
-}
-
-const ToastContext = createContext<ToastContextType | undefined>(undefined);
+import { useState, useCallback, type ReactNode } from 'react';
+import { ToastContext, type Toast } from './toast-context';
 
 export function ToastProvider({ children }: { children: ReactNode }) {
   const [toasts, setToasts] = useState<Toast[]>([]);
@@ -23,7 +7,7 @@ export function ToastProvider({ children }: { children: ReactNode }) {
   const toast = useCallback(({ title, description, variant = 'default' }: Omit<Toast, 'id'>) => {
     const id = Math.random().toString(36).substring(2, 9);
     setToasts((prev) => [...prev, { id, title, description, variant }]);
-    
+
     // Auto-dismiss after 5 seconds
     setTimeout(() => {
       setToasts((prev) => prev.filter((t) => t.id !== id));
@@ -39,12 +23,4 @@ export function ToastProvider({ children }: { children: ReactNode }) {
       {children}
     </ToastContext.Provider>
   );
-}
-
-export function useToast() {
-  const context = useContext(ToastContext);
-  if (!context) {
-    throw new Error('useToast must be used within a ToastProvider');
-  }
-  return context;
 }
