+++ projects/web/src/hooks/toast-context.ts
@@ -1,0 +1,18 @@
+import { createContext } from 'react';
+
+export type ToastVariant = 'default' | 'destructive' | 'success';
+
+export interface Toast {
+  id: string;
+  title?: string;
+  description?: string;
+  variant?: ToastVariant;
+}
+
+export interface ToastContextType {
+  toasts: Toast[];
+  toast: (props: Omit<Toast, 'id'>) => void;
+  dismiss: (id: string) => void;
+}
+
+export const ToastContext = createContext<ToastContextType | undefined>(undefined);
