+++ projects/web/src/hooks/use-toast.ts
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { ToastContext } from './toast-context';
+
+export type { Toast, ToastVariant, ToastContextType } from './toast-context';
+
+export function useToast() {
+  const context = useContext(ToastContext);
+  if (!context) {
+    throw new Error('useToast must be used within a ToastProvider');
+  }
+  return context;
+}
