+++ projects/web/src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import './index.css'
 import App from './App.tsx'
-import { ToastProvider } from './hooks/use-toast.tsx'
+import { ToastProvider } from './hooks/ToastProvider.tsx'
 import { Toaster } from './components/ui/toaster.tsx'
 
 createRoot(document.getElementById('root')!).render(
